Inline Wan VAE parallel helpers
