+++ python/sglang/multimodal_gen/runtime/models/vaes/ltx_2_vae.py
@@ -21,7 +21,7 @@ from sglang.multimodal_gen.runtime.distributed.parallel_state import (
     get_decode_parallel_world_size,
 )
 from sglang.multimodal_gen.runtime.models.vaes.common import ParallelTiledVAE
-from sglang.multimodal_gen.runtime.layers.spatial_parallel import (
+from sglang.multimodal_gen.runtime.layers.parallel_conv import (
     SpatialParallelConv3d,
     disable_spatial_parallel_decode,
     gather_and_trim_height,
