+++ python/sglang/multimodal_gen/test/unit/test_vae_loader.py
@@ -17,7 +17,7 @@ from sglang.multimodal_gen.runtime.loader.component_loaders.vae_loader import (
     _backfill_ltx2_audio_vae_latent_stats,
     _should_use_channels_last_3d,
 )
-from sglang.multimodal_gen.runtime.models.vaes.parallel import wan_common_utils
+from sglang.multimodal_gen.runtime.models.vaes import wanvae
 
 
 class _FakeServerArgs:
@@ -177,13 +177,13 @@ class TestVAELoader(unittest.TestCase):
 
         with (
             patch.object(
-                wan_common_utils.current_platform, "is_cuda", return_value=False
+                wanvae.current_platform, "is_cuda", return_value=False
             ),
             patch.object(
-                wan_common_utils.current_platform, "is_rocm", return_value=False
+                wanvae.current_platform, "is_rocm", return_value=False
             ),
         ):
-            out = wan_common_utils.match_conv3d_input_format(x, weight)
+            out = wanvae.match_conv3d_input_format(x, weight)
 
         self.assertIs(out, x)
 
@@ -198,13 +198,13 @@ class TestVAELoader(unittest.TestCase):
 
         with (
             patch.object(
-                wan_common_utils.current_platform, "is_cuda", return_value=True
+                wanvae.current_platform, "is_cuda", return_value=True
             ),
             patch.object(
-                wan_common_utils.current_platform, "is_rocm", return_value=False
+                wanvae.current_platform, "is_rocm", return_value=False
             ),
         ):
-            out = wan_common_utils.match_conv3d_input_format(x, weight)
+            out = wanvae.match_conv3d_input_format(x, weight)
 
         self.assertTrue(out.is_contiguous(memory_format=torch.channels_last_3d))
