+++ python/sglang/multimodal_gen/test/unit/test_vae_spatial_parallel_decode.py
@@ -41,18 +41,18 @@ from sglang.multimodal_gen.runtime.models.vaes.ltx_2_vae import (
     LTX2VideoDecoder3d,
     _enable_ltx_decoder_spatial_parallel,
 )
-from sglang.multimodal_gen.runtime.layers.spatial_parallel import (
+from sglang.multimodal_gen.runtime.layers.parallel_conv import (
     SpatialParallelCausalConv3d,
     SpatialParallelConv2d,
     SpatialParallelConv3d,
     chunk_height_by_sizes,
     split_for_parallel_decode,
 )
-from sglang.multimodal_gen.runtime.models.vaes.parallel.wan_dist_utils import (
+from sglang.multimodal_gen.runtime.models.vaes.wanvae import (
+    WanDecoder3d,
     WanDistAttentionBlock,
     WanDistCausalConv3d,
 )
-from sglang.multimodal_gen.runtime.models.vaes.wanvae import WanDecoder3d
 from sglang.multimodal_gen.runtime.utils.distributed import RankGenerator
 from sglang.multimodal_gen.utils import FlexibleArgumentParser
 
@@ -260,12 +260,12 @@ class TestVAESpatialParallelDecode(unittest.TestCase):
         x = torch.arange(10).view(1, 1, 10, 1)
 
         with patch(
-            "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_rank",
+            "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_rank",
             return_value=0,
         ):
             rank0 = chunk_height_by_sizes(x, [6, 4])
         with patch(
-            "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_rank",
+            "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_rank",
             return_value=1,
         ):
             rank1 = chunk_height_by_sizes(x, [6, 4])
@@ -289,11 +289,11 @@ class TestVAESpatialParallelDecode(unittest.TestCase):
                 return_value=0,
             ),
             patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_world_size",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_world_size",
                 return_value=2,
             ),
             patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_rank",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_rank",
                 return_value=0,
             ),
         ):
@@ -333,15 +333,11 @@ class TestVAESpatialParallelDecode(unittest.TestCase):
                 return_value=0,
             ),
             patch(
-                "sglang.multimodal_gen.runtime.models.vaes.parallel.wan_dist_utils.get_decode_parallel_world_size",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_world_size",
                 return_value=2,
             ),
             patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_world_size",
-                return_value=2,
-            ),
-            patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_rank",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_rank",
                 return_value=0,
             ),
         ):
@@ -391,11 +387,11 @@ class TestVAESpatialParallelDecode(unittest.TestCase):
                 return_value=2,
             ),
             patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_world_size",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_world_size",
                 return_value=2,
             ),
             patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_rank",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_rank",
                 return_value=0,
             ),
         ):
@@ -409,11 +405,11 @@ class TestVAESpatialParallelDecode(unittest.TestCase):
     def test_ltx_decoder_uses_spatial_parallel_conv3d(self):
         with (
             patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_world_size",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_world_size",
                 return_value=2,
             ),
             patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_rank",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_rank",
                 return_value=0,
             ),
         ):
@@ -441,11 +437,11 @@ class TestVAESpatialParallelDecode(unittest.TestCase):
     def test_hunyuan_decoder_uses_spatial_parallel_components(self):
         with (
             patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_world_size",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_world_size",
                 return_value=2,
             ),
             patch(
-                "sglang.multimodal_gen.runtime.layers.spatial_parallel.get_decode_parallel_rank",
+                "sglang.multimodal_gen.runtime.layers.parallel_conv.get_decode_parallel_rank",
                 return_value=0,
             ),
         ):
